Potato Wood Cauldron Nerf
The potato recipe in the wood cauldron can currently produce 1.5x as much xp/hour as the carrot recipe.  Rather than buffing the carrot recipe to bring it in line with the potato 1, it's better to nerf the potato recipe as the wood cauldron is already OP due to bots.