Made Pylon Locator Much Cheaper
After feedback from both community and the balance discussion Slack, it is agreed that the current cost of making something to detect the location of pylons is too expensive. A cheaper way to locate pylons should be tested at least, and we can make a final decision once this plays out.