Endgame EXP and pylon rebalancing (#291)
* Pylon balancing

Increase pylon runtime to a week

* Ruby Enrichment balances

Our town runs this recipe and we have a lot of data. The fish
requirement is out of whack with the pufferfish in the Topaz recipe.
Pufferfish are about half the chance of being caught of salmon. So it
seems fair that salmon amounts are doubled, not quadroupled.

Sapplings require really huge farms to get large amounts. It seems fair
to halve sappling amounts on all recipes.

For the rest this recipe seems fine. Vines cost iron to farm but this
can be mitigated by unbreaking shears.

* Topaz Rebalance

I don't have as much data on this recipe. But eggs take a long time to
get. Halving it seems fair.

Since I halved the sapplings on all other recipes I did it here as well.
But maybe 16 sapplings is too low? I don't really know how many saplings
dark oak trees drop in comparison.

* Amethys Balances

This recipe is worthless IMO. No one is going to take this path. It
requires 3 types of tree farms acacia, dark oak/oak for apples and
jungle. On top of that it takes ink.

I've reduced the amounts of hard to get items like apples and ink. But I
don't think it's enough to make this viable. Either oak or ink needs to
be changed to an easier to farm ingredient. Ink is already needed for
bastions as well, so I suggest that part be changed.

* Pylon rebalance

Changed to run for 2 days without supervision.

* Grandmaster recipes

This should bring all the recipes more in line with eachother. The fish
are based of the %chance to catch them, and each should take about the
same amount of effort. Clownfish 3, salmon 5 and fish 15.

* Changed clownfish to pufferfish in config